sc-14094 fix admin ui errors
