+++ web/gds-admin-ui/src/features/vasps/constants/query-key.ts
@@ -1,1 +1,3 @@
 export const GET_REVIEW_NOTES = 'get-review-notes';
+export const UPDATE_CONTACT = 'update-contact';
+export const DELETE_CONTACT = 'delete-contact';
