+++ web/gds-admin-ui/src/features/vasps/types/Contact.ts
@@ -1,0 +1,7 @@
+export type Contact = {
+    email: string;
+    extra: string;
+    name: string;
+    person: string;
+    phone: string;
+};
